requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/base.txt requirements/base.in
#
babel==2.5.1              # via django-phonenumber-field
cairocffi==0.8.0          # via cairosvg, weasyprint
cairosvg==2.1.1           # via weasyprint
certifi==2017.7.27.1      # via requests
cffi==1.11.2              # via cairocffi, weasyprint
chardet==3.0.4            # via requests
cssselect2==0.2.1         # via cairosvg
cssselect==1.0.1          # via weasyprint
defusedxml==0.5.0         # via cairosvg, python3-openid
diff-match-patch==20121119  # via django-import-export
django-admin-index==0.9.0
django-allauth==0.34.0
django-appconf==1.0.2
django-autoslug==1.9.3
django-axes==2.3.3
django-choices==1.6.0
django-ckeditor==5.4.0
django-compat==1.0.14     # via django-hijack, django-hijack-admin
django-countries==5.0
django-extensions==1.9.7
django-hijack-admin==2.1.4
django-hijack==2.1.4
django-import-export==0.5.1
django-js-asset==1.0.0    # via django-ckeditor
django-ordered-model==1.4.3  # via django-admin-index
django-phonenumber-field==1.3.0
django-redis==4.8.0
django-rosetta==0.7.13
django-sendfile==0.3.11
django-sniplates==0.5.0
django-solo==1.1.2
django==1.11.6
djangorestframework==3.7.3
et-xmlfile==1.0.1         # via openpyxl
html5lib==0.999999999     # via weasyprint
idna==2.6                 # via requests
jdcal==1.3                # via openpyxl
lxml==4.1.1               # via weasyprint
microsofttranslator==0.8  # via django-rosetta
oauthlib==2.0.6           # via requests-oauthlib
odfpy==1.3.5              # via tablib
olefile==0.44             # via pillow
openpyxl==2.4.9           # via tablib
phonenumberslite==8.8.3   # via django-phonenumber-field
pillow==4.3.0
polib==1.0.8              # via django-rosetta
psycopg2==2.7.3.1
pycparser==2.18           # via cffi
pyphen==0.9.4             # via weasyprint
python3-openid==3.1.0     # via django-allauth
pytz==2017.2
pyyaml==3.12              # via tablib
raven==6.2.1
redis==2.10.6             # via django-redis
requests-oauthlib==0.8.0  # via django-allauth
requests==2.18.4          # via django-allauth, django-rosetta, microsofttranslator, requests-oauthlib
rules==1.2.1
six==1.11.0               # via django-compat, django-extensions, django-rosetta, html5lib, microsofttranslator
tablib==0.12.1            # via django-import-export
tinycss2==0.6.1           # via cairosvg, cssselect2, weasyprint
typing==3.6.2             # via django-extensions
unicodecsv==0.14.1        # via tablib
urllib3==1.22             # via requests
weasyprint==0.38
webencodings==0.5.1       # via html5lib, tinycss2
xlrd==1.1.0               # via tablib
xlwt==1.3.0               # via tablib
